+++ components/logins/src/error.rs
@@ -62,6 +62,12 @@ pub enum LoginsApiError {
 /// is never returned to the consumer.
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
+    #[error("Database is closed")]
+    DatabaseClosed,
+
+    #[error("EncryptorDecryptor is closed")]
+    EncryptorDecryptorClosed,
+
     #[error("Malformed incoming record")]
     MalformedIncomingRecord,